requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
boto3==1.21.25
botocore==1.24.25
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
idna==3.3
jmespath==1.0.0
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8